src/requirements.txt
beautifulsoup4==4.6.0
certifi==2020.6.20
chardet==3.0.4
deprecation==2.1.0
eyeD3==0.9.5
filetype==1.0.7
idna==2.10
lyricsgenius==1.8.6
packaging==20.4
pyparsing==2.4.7
python-mpd2==1.1.0
requests==2.24.0
six==1.15.0
urllib3==1.25.9
